PTASK CAB library updated with a timestamp, audio recording now uses "oversampling"
The "Oversampling" basically will ask at a higher rate wether the capture buffer contains some data or not and if so it will keep filling its own buffer, until it's full.
This of course means that some executions will take less if no time to be performed, but data in general will be available as soon as possible to be analyzed.